requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.6
django-filter==24.3
django-rest-swagger==2.2.0
djangorestframework==3.15.2
drf-spectacular==0.28.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.5
Markdown==3.7
MarkupSafe==3.0.2
openapi-codec==1.3.2
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
